Implement hours&minutes functionality PART-1 (doesn't work)
- The time column currently shows duplicated time

+++ core/src/main/java/com/alamkanak/weekview/Renderers.kt
@@ -17,7 +17,7 @@ internal interface DateFormatterDependent {
     fun onDateFormatterChanged(formatter: DateFormatter)
 }
 
-typealias TimeFormatter = (Int) -> String
+typealias TimeFormatter = (Double) -> String
 
 internal interface TimeFormatterDependent {
     fun onTimeFormatterChanged(formatter: TimeFormatter)
